% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{add_internal_numbers}
\alias{add_internal_numbers}
\title{Only for internal number adding}
\usage{
add_internal_numbers(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\description{
Only for internal number adding
}
