% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{add_numbers}
\alias{add_numbers}
\title{Add together two numbers.}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
the The sum of \code{x} and \code{y}
}
\description{
Add together two numbers.
}
\examples{
add_numbers(1, 2) # add 1 + 2
}
